Simplify repo to minimal odds API demo
